feat: decoder para sensor sem dente de falha (distribuidor e volante de 1 dente)
Suporta sensores em que cada pulso ja e uma referencia angular, sem gap para
procurar. Cobre dois casos que na pratica sao o mesmo problema:

- distribuidor com sensor hall (Opala): um pulso por cilindro, sensor girando
  na rotacao do comando
- volante com um unico dente (motos, tipo Titan): um pulso por volta do
  virabrequim

Selecionado por tipo_ignicao = 2, campo que ja existia na configuracao com esse
significado documentado mas nunca era consultado por nenhuma logica. Sem bloco
novo de protocolo, sem mudanca de EEPROM, sem acoplamento com a UI.

Este decoder e mais SIMPLES que o de roda fonica, nao mais complexo: sem
deteccao de gap, sem contagem de dentes, sem validacao de sincronismo. Em troca
nao ha como confirmar sincronismo - um pulso espurio vira um evento espurio, e
a unica defesa e o filtro de ruido, reaproveitado com o mesmo escape contra
travamento permanente.

Mudancas:
- decoder.h: decoder_sem_falha() e despacho em leitor_sensor_roda_fonica
- util.h: sensor_sem_falha(); offset_referencia_roda_fonica_graus retorna 0
  neste modo, porque o angulo e medido a partir do proprio pulso
- ignicao.h / injecao.h: 1 evento por pulso, sempre na saida 0 (o distribuidor
  roteia mecanicamente; na moto ha uma bobina so, eventualmente dupla em
  paralelo para gemeas - centelha perdida cobre os dois)
- timer.h: ticks_entre_referencias() substitui o "360 * tempo_cada_grau" fixo
  do cruzamento de referencia. Ver abaixo.
- variaveis_global.h: grau_cada_dente de byte para int

grau_cada_dente precisava de int porque com 1 dente vale 360, e um byte
truncaria para 104 - o angulo sairia deslocado em quase 3,5x. Mesma classe dos
bugs de grau_entre_cada_cilindro (58c312b) e do dreq_fuel (4b28973).

O primeiro teste de bancada pegou um erro que so aparece neste modo: a logica
de empurrar o evento para a referencia seguinte somava 360 graus fixos, ou seja
uma volta INTEIRA do sensor. Sem dente de falha a referencia se repete a cada
grau_cada_dente graus (60 num distribuidor de 6 cilindros), entao o evento era
jogado 6 pulsos a frente e o canal ficava armado no meio tempo, pulando os
agendamentos seguintes. Media 0,17 eventos por pulso - exatamente 1/6, com o
angulo ja correto. Corrigido com ticks_entre_referencias(), que devolve o
periodo certo para cada modo.

Validado em bancada nas duas configuracoes:

  DISTRIBUIDOR  g,2,6,1,0,30,6   sensor 499rpm (motor ~999rpm)
    ign1   1,00 por pulso, 328/328 pulsos com centelha
    INJ1   1,00 por pulso, abertura em 25,2 graus (grau_pms 30 - fechamento 5)
    ign2 e ign3 mudos

  VOLANTE 1 DENTE  g,2,1,2,0,30,6   motor 2996rpm
    ign1   1,00 por volta, 358/358 voltas com centelha
           centelha em 3,5 graus apos o pulso -> avanco efetivo 26,5 graus
           APMS, coerente com a tabela na faixa de 3000rpm
    INJ1   1,00 por volta, abertura em 27,2 graus
    ign2 e ign3 mudos

Limitacao inerente ao sensor, nao ao firmware: com 1 dente o firmware
extrapola 360 graus inteiros a partir do intervalo anterior, entao a dispersao
angular e visivelmente maior que a do distribuidor. Documentado no README junto
com a tabela de configuracao dos dois casos.

+++ include/variaveis_global.h
@@ -33,7 +33,9 @@ static inline unsigned int normalizar_dwell_us(long valor) {
 
 byte tipo_ignicao_sequencial = 0;// sequencial 1 semi-sequencial 0
 volatile unsigned int qtd_voltas = 0;
-byte grau_cada_dente = 360 / qtd_dente;
+// int, nao byte: com um unico dente no volante (moto) o valor e 360 e um byte
+// truncaria para 104. Mesma classe do bug de grau_entre_cada_cilindro.
+int grau_cada_dente = 360 / qtd_dente;
 byte grau_avanco = 0;
 byte grau_avanco_partida = 1; // avanço definido apenas na partida
 // Com sensor no virabrequinho, o ciclo completo de combustao dura 720 graus
